Arbiter configuration reload - tested and enhanced (reload_configuration http endpoint) Arbiter configuration reload - reload several successive times rather than only once if several requests are raised
