+++ alignak/misc/common.py
@@ -43,7 +43,9 @@
 This module is used for common variables in Alignak.
 Previously some of those variables were linked to a specific class which made no sense.
 """
+import signal
 from collections import namedtuple
+
 try:
     from setproctitle import setproctitle  # pylint: disable=unused-import
 except ImportError as err:  # pragma: no cover, setproctitle is in the requirements.txt
@@ -55,6 +57,10 @@ except ImportError as err:  # pragma: no cover, setproctitle is in the requireme
         :return: None
         """
         return None
+
+# Friendly names for the system signals
+SIGNALS_TO_NAMES_DICT = dict((k, v) for v, k in reversed(sorted(signal.__dict__.items()))
+                             if v.startswith('SIG') and not v.startswith('SIG_'))
 
 ModAttr = namedtuple('ModAttr', ['modattr', 'attribute', 'value'])
