+++ test/cfg/run_passive/arbiter/objects/commands/detailled-service-by-email.cfg
@@ -1,0 +1,7 @@
+
+## Notify Service by Email with detailled informations
+# Service have appropriate macros. Look at unix-fs pack to get an example
+define command {
+    command_name    detailled-service-by-email
+    command_line    /usr/bin/printf "%b" "Alignak Notification\n\nNotification Type: $NOTIFICATIONTYPE$\n\nService: $SERVICEDESC$\nHost: $HOSTALIAS$\nAddress: $HOSTADDRESS$\nState: $SERVICESTATE$\n\nDate/Time: $DATE$ at $TIME$\nService Output : $SERVICEOUTPUT$\n\nService Description: $_SERVICEDETAILLEDESC$\nService Impact: $_SERVICEIMPACT$\nFix actions: $_SERVICEFIXACTIONS$" | /usr/bin/mail -s "$SERVICESTATE$ on Host : $HOSTALIAS$/Service : $SERVICEDESC$" $CONTACTEMAIL$
+}
