+++ test/cfg/run_passive/arbiter/objects/commands/dummy_check.cfg
@@ -1,0 +1,6 @@
+## Notify Host by Email with detailled informations
+# Service have appropriate macros. Look at unix-fs pack to get an example
+define command {
+    command_name    dummy_check
+    command_line    /tmp/dummy_command.sh $ARG1$ $ARG2$
+}
