+++ test/cfg/run_passive/arbiter/objects/commands/notify-host-by-email.cfg
@@ -1,0 +1,5 @@
+## Notify Host by Email
+define command {
+    command_name    notify-host-by-email
+    command_line    /usr/bin/printf "%b" "Alignak Notification\n\nType:$NOTIFICATIONTYPE$\nHost: $HOSTNAME$\nState: $HOSTSTATE$\nAddress: $HOSTADDRESS$\nInfo: $HOSTOUTPUT$\nDate/Time: $DATE$ $TIME$\n" | /usr/bin/mail -s "Host $HOSTSTATE$ alert for $HOSTNAME$" $CONTACTEMAIL$
+}
