+++ test/cfg/run_passive/arbiter/objects/commands/notify-service-by-email.cfg
@@ -1,0 +1,6 @@
+## Notify Service by Email
+define command {
+    command_name    notify-service-by-email
+    command_line    /usr/bin/printf "%b" "Alignak Notification\n\nNotification Type: $NOTIFICATIONTYPE$\n\nNotification number: $SERVICENOTIFICATIONNUMBER$\n\nService: $SERVICEDESC$\nHost: $HOSTNAME$\nAddress: $HOSTADDRESS$\nState: $SERVICESTATE$\n\nDate/Time: $DATE$ $TIME$\nAdditional Info : $SERVICEOUTPUT$\n" | /usr/bin/mail -s "** $NOTIFICATIONTYPE$ alert - $HOSTNAME$/$SERVICEDESC$ is $SERVICESTATE$ **" $CONTACTEMAIL$
+}
+
