+++ test/cfg/run_passive/arbiter/objects/contactgroups/admins.cfg
@@ -1,0 +1,5 @@
+define contactgroup{
+    contactgroup_name   admins
+    alias               Administrators
+    members             admin
+}
