+++ test/cfg/run_passive/arbiter/objects/contactgroups/users.cfg
@@ -1,0 +1,5 @@
+define contactgroup{
+    contactgroup_name   users
+    alias               Guest users
+    members             guest
+}
