+++ test/cfg/run_passive/arbiter/objects/contacts/admin.cfg
@@ -1,0 +1,11 @@
+define contact{
+    use                 generic-contact
+    contact_name        admin
+    alias               Administrator
+    email               frederic.mohier@alignak.net
+    pager               0600000000   ; contact phone number
+    password            admin
+    is_admin            1
+    ;can_submit_commands     1 (implicit because is_admin)
+}
+
