+++ test/cfg/run_passive/arbiter/objects/contacts/guest.cfg
@@ -1,0 +1,9 @@
+define contact{
+    use                 generic-contact
+    contact_name        guest
+    alias               Guest
+    email               guest@localhost
+    password            guest
+    is_admin                0
+    can_submit_commands 0
+}
