+++ test/cfg/run_passive/arbiter/objects/hosts/localhost.cfg
@@ -1,0 +1,14 @@
+define host{
+    use                     generic-host
+    contact_groups          admins
+    host_name               localhost
+    alias                   Web UI
+    display_name            Alignak Web UI
+    address                 127.0.0.1
+
+    hostgroups              monitoring_servers
+
+    # Web UI host importance
+    # Business impact (from 0 to 5)
+    business_impact          4
+}
