+++ test/cfg/run_passive/arbiter/objects/notificationways/detailled-email.cfg
@@ -1,0 +1,12 @@
+# This is how emails are sent, 24x7 way.
+define notificationway{
+   notificationway_name            detailled-email
+   service_notification_period     24x7
+   host_notification_period        24x7
+   service_notification_options    c,w,r
+   host_notification_options       d,u,r,f,s
+   service_notification_commands   detailled-service-by-email ; send service notifications via email
+   host_notification_commands      detailled-host-by-email    ; send host notifications via email
+   min_business_impact             1
+}
+
