+++ test/cfg/run_passive/arbiter/objects/notificationways/email.cfg
@@ -1,0 +1,11 @@
+# This is how emails are sent, 24x7 way.
+define notificationway{
+       notificationway_name            email
+       service_notification_period     24x7
+       host_notification_period        24x7
+       service_notification_options    c,w,r
+       host_notification_options       d,u,r,f,s
+       service_notification_commands   notify-service-by-email ; send service notifications via email
+       host_notification_commands      notify-host-by-email    ; send host notifications via email
+}
+
