+++ test/cfg/run_passive/arbiter/objects/timeperiods/24x7.cfg
@@ -1,0 +1,12 @@
+define timeperiod{
+	timeperiod_name			24x7
+	alias				Always
+	sunday				00:00-24:00
+	monday				00:00-24:00
+	tuesday				00:00-24:00
+	wednesday			00:00-24:00
+	thursday			00:00-24:00
+	friday				00:00-24:00
+	saturday			00:00-24:00
+}
+
