+++ test/cfg/run_passive/arbiter/objects/timeperiods/none.cfg
@@ -1,0 +1,5 @@
+# 'none' timeperiod definition
+define timeperiod{
+	timeperiod_name	none
+	alias		No Time Is A Good Time
+	}
