+++ test/cfg/run_passive/arbiter/objects/timeperiods/us-holidays.cfg
@@ -1,0 +1,16 @@
+# Some U.S. holidays
+# Note: The timeranges for each holiday are meant to *exclude* the holidays from being
+# treated as a valid time for notifications, etc.  You probably don't want your pager
+# going off on New Year's.  Although you're employer might... :-)
+define timeperiod{
+	name			us-holidays
+        timeperiod_name         us-holidays
+        alias                   U.S. Holidays
+
+        january 1               00:00-00:00     ; New Years
+        monday -1 may           00:00-00:00     ; Memorial Day (last Monday in May)
+        july 4                  00:00-00:00     ; Independence Day
+        monday 1 september      00:00-00:00     ; Labor Day (first Monday in September)
+        thursday -1 november    00:00-00:00     ; Thanksgiving (last Thursday in November)
+        december 25             00:00-00:00     ; Christmas
+        }
