+++ test/cfg/run_passive/arbiter/objects/timeperiods/workhours.cfg
@@ -1,0 +1,10 @@
+# 'workhours' timeperiod definition
+define timeperiod{
+	timeperiod_name	workhours
+	alias		Normal Work Hours
+	monday		09:00-17:00
+	tuesday		09:00-17:00
+	wednesday	09:00-17:00
+	thursday	09:00-17:00
+	friday		09:00-17:00
+	}
