+++ test/cfg/run_passive/arbiter/realms/All/hosts.cfg
@@ -1,0 +1,10 @@
+define host{
+    use                     generic-host
+    contact_groups          admins
+    host_name               alignak-all-00
+    alias                   Alignak
+    display_name            Alignak (Demo)
+    address                 127.0.0.1
+
+    check_command           dummy_check!0
+}
