+++ test/cfg/run_passive/arbiter/realms/All/realm.cfg
@@ -1,0 +1,6 @@
+# Very advanced feature for multisite management.
+# Read the docs VERY CAREFULLY before changing these settings :)
+define realm {
+    realm_name      All
+    default         1
+}
