+++ test/cfg/run_passive/arbiter/realms/All/services.cfg
@@ -1,0 +1,36 @@
+define service{
+  check_command                  _echo
+  host_name                      alignak-all-00
+  service_description            dummy_echo
+  use                            generic-service
+}
+define service{
+  check_command                  dummy_check!0
+  host_name                      alignak-all-00
+  service_description            dummy_ok
+  use                            generic-service
+}
+define service{
+  check_command                  dummy_check!1
+  host_name                      alignak-all-00
+  service_description            dummy_warning
+  use                            generic-service
+}
+define service{
+  check_command                  dummy_check!2
+  host_name                      alignak-all-00
+  service_description            dummy_critical
+  use                            generic-service
+}
+define service{
+  check_command                  dummy_check
+  host_name                      alignak-all-00
+  service_description            dummy_unknown
+  use                            generic-service
+}
+define service{
+  check_command                  dummy_check!0!10
+  host_name                      alignak-all-00
+  service_description            dummy_timeout
+  use                            generic-service
+}
