+++ test/cfg/run_passive/arbiter/resource.d/paths.cfg
@@ -1,0 +1,7 @@
+# Nagios legacy macros
+$USER1$=$NAGIOSPLUGINSDIR$
+$NAGIOSPLUGINSDIR$=/usr/lib/nagios/plugins
+
+#-- Location of the plugins for Alignak
+$PLUGINSDIR$=/tmp/var/libexec/alignak
+
