+++ test/cfg/run_passive/arbiter/templates/business-impacts.cfg
@@ -1,0 +1,81 @@
+# Some business impact templates
+# ------------------------------
+# The default value for business impact is 2, meaning "normal".
+
+define host{
+    register                0
+    name                    no-importance
+    business_impact         0
+    # Disable notifications
+    notifications_enabled   0
+}
+
+define host{
+    register                0
+    name                    qualification
+    business_impact         1
+}
+
+define host{
+    register                0
+    name                    normal
+    business_impact         2
+}
+
+define host{
+    register                0
+    name                    production
+    business_impact         3
+}
+
+define host{
+    register                0
+    name                    important
+    business_impact         4
+}
+
+define host{
+    register                0
+    name                    top-for-business
+    business_impact         5
+}
+
+
+define service{
+    register                0
+    name                    no-importance
+    business_impact         0
+    # Disable notifications
+    notifications_enabled   0
+}
+
+define service{
+    register                0
+    name                    qualification
+    business_impact         1
+}
+
+define service{
+    register                0
+    name                    normal
+    business_impact         2
+}
+
+define service{
+    register                0
+    name                    production
+    business_impact         3
+}
+
+define service{
+    register                0
+    name                    important
+    business_impact         4
+}
+
+define service{
+    register                0
+    name                    top-for-business
+    business_impact         5
+}
+
