+++ test/cfg/run_passive/arbiter/templates/generic-contact.cfg
@@ -1,0 +1,11 @@
+# Contact definition
+# By default the contact will ask notification by mails
+define contact{
+        name                            generic-contact
+	host_notifications_enabled	1
+	service_notifications_enabled	1
+	email				alignak@localhost
+	can_submit_commands		1
+	notificationways        	email
+        register                        0
+	}
