+++ test/cfg/run_passive/arbiter/templates/generic-host.cfg
@@ -1,0 +1,42 @@
+# Generic host definition template - This is NOT a real host, just a template!
+# Most hosts should inherit from this one
+define host{
+    name                    generic-host
+
+    # Checking part
+    check_command           _internal_host_up
+    max_check_attempts      2
+    check_interval          1
+
+    # Check every time
+    active_checks_enabled   1
+    check_period            24x7
+
+    # Notification part
+    # One notification each day (1440 = 60min* 24h)
+    # every time, and for all 'errors'
+    # notify the admins contactgroups by default
+    contact_groups          admins,users
+    notification_interval   1440
+    notification_period     24x7
+    notification_options    d,u,r,f
+    notifications_enabled   1
+
+    # Advanced option
+    event_handler_enabled   0
+    flap_detection_enabled  1
+    process_perf_data       1
+    snapshot_enabled        0
+
+    # Maintenance / snapshot period
+    #maintenance_period      none
+    #snapshot_period         none
+
+    # Dispatching
+    #poller_tag          DMZ
+    #realm               All
+
+    # This to say that it's a template
+    register            0
+}
+
