+++ test/cfg/run_passive/arbiter/templates/generic-service.cfg
@@ -1,0 +1,20 @@
+# Generic service definition template - This is NOT a real service, just a template!
+define service{
+    name                            generic-service         ; The 'name' of this service template
+    active_checks_enabled           1                       ; Active service checks are enabled
+    passive_checks_enabled          1                       ; Passive service checks are enabled/accepted
+    notifications_enabled           1                       ; Service notifications are enabled
+    notification_interval           1
+    notification_period             24x7
+    event_handler_enabled           0                       ; Service event handler is enabled
+    flap_detection_enabled          1                       ; Flap detection is enabled
+    process_perf_data               1                       ; Process performance data
+    is_volatile                     0                       ; The service is not volatile
+    check_period                    24x7                    ; The service can be checked at any time of the day
+    max_check_attempts              1                       ; Re-check the service only once to determine its final (hard) state
+    check_interval                  1                       ; Check the service every minute under normal conditions
+    retry_interval                  1                       ; Re-check the service every minute until a hard state can be determined
+    notification_options            w,u,c,r                 ; Send notifications about warning, unknown, critical, and recovery events
+    contact_groups                  admins,users
+    register                        0                       ; DONT REGISTER THIS DEFINITION - ITS NOT A REAL SERVICE, JUST A TEMPLATE
+}
